+++ Projects/App/Sources/Service/OctopusClientsImplementation/CrossSellClientOctopus.swift
@@ -3,16 +3,6 @@ import Foundation
 import hCore
 import hGraphQL
 
-@MainActor
-public class CrossSellService {
-    @Inject var service: CrossSellClient
-
-    public func getCrossSell() async throws -> [CrossSell] {
-        log.info("CrossSellService: getCrossSell", error: nil, attributes: nil)
-        return try await service.getCrossSell()
-    }
-}
-
 public class CrossSellClientOctopus: CrossSellClient {
     @Inject private var octopus: hOctopus
     public init() {}
